Changed timestart from float to int.

+++ flink/flink-python-job/src/main/java/org/lmu/JobNameBatchDB.java
@@ -36,7 +36,7 @@ public final class JobNameBatchDB {
         JSONObject bestResultJosnObject = flinkdistribute.getBestMapeJsonObject(resCollect);
         System.out.println("Save best result with status Finished");
         bestResultJosnObject.put("status", "Finished");
-        bestResultJosnObject.put("timestart", jobsjson.get("timestart"));
+        bestResultJosnObject.put("timestart", (int) jobsjson.get("timestart"));
         bestResultJosnObject.put("timeend", System.currentTimeMillis());
         flinkdistribute.saveResultJSONObjectToMongoDB(bestResultJosnObject);
     }
